requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
python-multipart==0.0.5
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0
cryptography==3.4.7
pyotp==2.6.0
qrcode==7.3
Pillow==8.3.1
geopy==2.2.0
scikit-learn==0.24.2
numpy==1.21.1
pandas==1.3.1
requests==2.26.0
pyjwt==2.1.0
python-dotenv==0.19.0
aiofiles==0.7.0
jinja2==3.0.1
itsdangerous==2.0.1
python-dateutil==2.8.2
pytz==2021.1
schedule==1.1.0
pycryptodome==3.10.1
